Additions and changes to support the OSG_Windows port of VESS.  Note that some files needed to be moved from common to OS-specific directories to support this port.

+++ graphics/OSG/vsDecalCallback.c++
@@ -72,11 +72,11 @@ void vsDecalCallback::operator()(osg::Node* node, osg::NodeVisitor* nv)
 
     // Make sure we have enough state sets to go around
     checkSize(decalGroup->getNumChildren());
-    if (stateSetArraySize < decalGroup->getNumChildren())
+    if (stateSetArraySize < (int)decalGroup->getNumChildren())
         return;
 
     // Run the cull traversal, by hand, on each of the Group's children
-    for (loop = 0; loop < decalGroup->getNumChildren(); loop++)
+    for (loop = 0; loop < (int)decalGroup->getNumChildren(); loop++)
     {
         // Push the polygon offset state set onto the CullVisitor's stack
         cullVisitor->pushStateSet((osg::StateSet *)(stateSetArray[loop]));
